Hid token, removed commands
Commands was causing cyclical dependency, so removed.
